requirements.txt
python-telegram-bot
requests
langgraph
python-dotenv
langgraph-cli[inmem]
langgraph-api
langchain
langchain_community
langchain-ollama
langchain-openai
duckduckgo_search
beautifulsoup4
certifi
langgraph-checkpoint-sqlite
aiosqlite
tabulate
pandas
matplotlib
telegramify_markdown
telegramify-markdown[mermaid]